requirements.txt
psycopg2-binary
Faker
pandas